+++ plugin/src/test/kotlin/org/jlleitschuh/gradle/ktlint/KtlintPluginVersionTest.kt
@@ -35,16 +35,16 @@ class KtlintPluginVersionTest : AbstractPluginTest() {
     }
 
     @Test
-    fun `with ktlint version equal to 0_20`() {
-        projectRoot.buildScriptUsingKtlintVersion("0.20.0")
+    fun `with ktlint version equal to 0_22`() {
+        projectRoot.buildScriptUsingKtlintVersion("0.22.0")
         build("ktlintCheck").apply {
             assertThat(task(":ktlintMainSourceSetCheck")!!.outcome, equalTo(TaskOutcome.SUCCESS))
         }
     }
 
     @Test
-    fun `with ktlint version less than 0_20`() {
-        projectRoot.buildScriptUsingKtlintVersion("0.19.0")
+    fun `with ktlint version less than 0_22`() {
+        projectRoot.buildScriptUsingKtlintVersion("0.21.0")
         buildAndFail("ktlintCheck").apply {
             assertThat(task(":ktlintMainSourceSetCheck")!!.outcome, equalTo(TaskOutcome.FAILED))
         }